# server/api.yml
openapi: "3.0.3"
info:
  title: TechMentorX API
  description: Backend API for TechMentorX student performance tracking, engagement analysis, and mental health support platform.
  version: "1.0.0"
servers:
  - url: http://localhost:5000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher]
        department:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            status:
              type: integer
...
paths:
...
  /auth/register:
    post:
      summary: Register a new student or teacher
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role, department]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
                role:
                  type: string
                  enum: [student, teacher]
                department:
                  type: string
                enrollmentId:
                  type: string
                  description: Required for students
                semester:
                  type: integer
                  description: Required for students
                employeeId:
                  type: string
                  description: Required for teachers
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      summary: Login with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /study/topics:
    get:
      summary: Get student's study topics ("To Study" section)
      tags: [Study]
      security:
        - BearerAuth: []
      description: "Role: student only"
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in-progress, completed]
        - name: subject
          in: query
          schema:
            type: string
          description: Subject ObjectId filter
      responses:
        "200":
          description: Study topics with summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      type: object
                  summary:
                    type: object
                    properties:
                      total:
                        type: integer
                      pending:
                        type: integer
                      inProgress:
                        type: integer
                      completed:
                        type: integer
                      overdue:
                        type: integer
  /study/topics/{topicId}:
    patch:
      summary: Update study topic status or progress
      tags: [Study]
      security:
        - BearerAuth: []
      description: "Role: student only"
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, in-progress, completed]
                aiProgress:
                  type: number
                  minimum: 0
                  maximum: 100
      responses:
        "200":
          description: Topic updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  topic:
                    type: object
  /study/topics/{topicId}/learn:
    post:
      summary: Get AI teaching content for a weak topic
      tags: [Study]
      security:
        - BearerAuth: []
      description: "Role: student only. Uses the lecture PPT to generate teaching content."
      parameters:
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: AI-generated teaching content
          content:
            application/json:
              schema:
                type: object
                properties:
                  topic:
                    type: string
                  content:
                    type: object
                    properties:
                      content:
                        type: string
                      exercises:
                        type: array
                        items:
                          type: object
                          properties:
                            question:
                              type: string
                            answer:
                              type: string
                      summary:
                        type: string
        "503":
          description: AI service unavailable
...